+++ OUDS/Core/Components/Sources/Controls/Checkbox/OUDSCheckboxItemIndeterminate.swift
@@ -166,61 +166,6 @@ public struct OUDSCheckboxItemIndeterminate: View {
     /// Creates a checkbox with label and optional helper text, icon, divider.
     ///
     /// ```swift
-    ///     OUDSCheckboxItemIndeterminate(selection: $state, label: "Select all")
-    /// ```
-    ///
-    /// **The design system does not allow to have both an error situation and a read only mode for the component.**
-    ///
-    /// - Parameters:
-    ///   - selection: A binding to a property that determines whether the indicator is ticked, unticked or preticked (indeterminate / partially ticked).
-    ///   - label: The main label text of the checkbox, must not be empty
-    ///   - description: A description, an additional helper text, should not be empty
-    ///   - icon: An optional icon
-    ///   - flipIcon: Default set to `false`, set to true to reverse the image (i.e. flip vertically)
-    ///   - isReversed: `true` if the checkbox indicator must be in trailing position, `false` otherwise. Default to `false`
-    ///   - isError: `true` if the look and feel of the component must reflect an error state, default set to `false`
-    ///   - errorText: An optional error message to display at the bottom. This message is ignored if `isError` is `false`.
-    ///   The `errorText`can be different if switch is selected or not.
-    ///   - isReadOnly: True if component is in read only, i.e. not really disabled but user cannot interact with it yet, default set to `false`
-    ///   - hasDivider: If `true` a divider is added at the bottom of the view, by default set to `false`
-    ///   - constrainedMaxWidth: When `true`, the item width is constrained to a maximum value defined by the design system.
-    ///     When `false`, no specific width constraint is applied, allowing the component to size itself or follow external
-    ///     modifier. Defaults to `false`.
-    ///   - action: An additional action to trigger when the checkbox has been pressed, default set to `nil`
-    ///
-    /// **Remark: If `label` and `description` strings are wording keys from strings catalog stored in `Bundle.main`, they are automatically localized. Else, prefer to
-    /// provide the localized string if key is stored in another bundle.**
-    @available(*, deprecated, message: "Use instead OUDSCheckboxItemIndeterminate(:selection:)")
-    public init(selection: Binding<OUDSCheckboxIndicatorState>,
-                label: String,
-                description: String? = nil,
-                icon: Image? = nil,
-                flipIcon: Bool = false,
-                isReversed: Bool = false,
-                isError: Bool = false,
-                errorText: String? = nil,
-                isReadOnly: Bool = false,
-                hasDivider: Bool = false,
-                constrainedMaxWidth: Bool = false,
-                action: (() -> Void)? = nil)
-    {
-        self.init(label,
-                  selection: selection,
-                  description: description,
-                  icon: icon,
-                  flipIcon: flipIcon,
-                  isReversed: isReversed,
-                  isError: isError,
-                  errorText: errorText,
-                  isReadOnly: isReadOnly,
-                  hasDivider: hasDivider,
-                  constrainedMaxWidth: constrainedMaxWidth,
-                  action: action)
-    }
-
-    /// Creates a checkbox with label and optional helper text, icon, divider.
-    ///
-    /// ```swift
     ///     OUDSCheckboxItemIndeterminate("Select all", selection: $state)
     /// ```
     ///