+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios/compare/1.4.0...develop)
 
+### Removed
+
+- **BREAKING**: Old methods prefixed by `ouds`
+- **BREAKING**: `UnorderedIcon` type
+- **BREAKING**: `OUDSBulletList.UnorderedAsset.free` case
+- **BREAKING**: `OUDSCheckboxItem(isOn:)`, `OUDSRadioItem(isOn:)`, `OUDSCheckboxItemIndeterminate(selection:)` inits
+- **BREAKING**: `OUDSTabBar(selected:count)` init
+
 ### Fixed
 
 - Overlay items of `tab bar` component still displayed even if hidden for disabled Liquid Glass (Orange-OpenSource/ouds-ios#1434)